+++ packages/bff/src/db.ts
@@ -1,8 +1,10 @@
 import { DataSource, type Repository } from 'typeorm';
-import { ProfileTable, SavedSearch } from './entities.ts';
+import { Group, Party, ProfileTable, SavedSearch } from './entities.ts';
 
+export let GroupRepository: Repository<Group> | undefined = undefined;
 export let ProfileRepository: Repository<ProfileTable> | undefined = undefined;
 export let SavedSearchRepository: Repository<SavedSearch> | undefined = undefined;
+export let PartyRepository: Repository<Party> | undefined = undefined;
 
 export let dataSource: DataSource | undefined = undefined;
 
@@ -12,9 +14,17 @@ export const connectToDB = async () => {
   const { connectionOptions } = await import('./data-source.ts');
   dataSource = await new DataSource(connectionOptions).initialize();
 
+  GroupRepository = dataSource.getRepository(Group);
   ProfileRepository = dataSource.getRepository(ProfileTable);
   SavedSearchRepository = dataSource.getRepository(SavedSearch);
+  PartyRepository = dataSource.getRepository(Party);
 
+  if (!GroupRepository) {
+    throw new Error('GroupRepository not initialized');
+  }
+  if (!PartyRepository) {
+    throw new Error('PartyRepository not initialized');
+  }
   if (!ProfileRepository) {
     throw new Error('ProfileRepository not initialized');
   }