+++ packages/bff/src/entities.ts
@@ -3,6 +3,9 @@ import {
   CreateDateColumn,
   DeleteDateColumn,
   Entity,
+  JoinColumn,
+  JoinTable,
+  ManyToMany,
   ManyToOne,
   OneToMany,
   PrimaryColumn,
@@ -18,11 +21,18 @@ export class ProfileTable {
   @Column({ length: 255, nullable: true })
   language: string;
 
-  @OneToMany('saved_search', 'profile')
+  @OneToMany(
+    () => SavedSearch,
+    (savedSearch) => savedSearch.profile,
+  )
   savedSearches: SavedSearch[];
 
-  @Column('text', { array: true })
-  favoriteActors: string[];
+  @OneToMany(
+    () => Group,
+    (group) => group.profile,
+    { cascade: false },
+  )
+  groups: Group[];
 
   @CreateDateColumn()
   createdAt: Date;
@@ -34,6 +44,39 @@ export class ProfileTable {
   deletedAt: Date;
 }
 
+@Entity({ name: 'group' })
+export class Group {
+  @PrimaryGeneratedColumn()
+  id: number;
+
+  @Column()
+  name: string;
+
+  @Column({ default: false })
+  isfavorite: boolean;
+
+  @ManyToOne(
+    () => ProfileTable,
+    (profile) => profile.groups,
+    { onDelete: 'NO ACTION' },
+  )
+  @JoinColumn({ name: 'profilePid', referencedColumnName: 'pid' })
+  profile: ProfileTable;
+
+  @ManyToMany('Party', (party: Party) => party.groups)
+  @JoinTable()
+  parties: Party[];
+}
+
+@Entity({ name: 'party' })
+export class Party {
+  @PrimaryColumn()
+  id: string;
+
+  @ManyToMany('Group', (group: Group) => group.parties)
+  groups: Group[];
+}
+
 export interface Filter {
   id: string;
   value: string;
@@ -43,6 +86,7 @@ export interface SavedSearchData {
   searchString?: string;
   fromView?: string;
 }
+
 @Entity({ name: 'saved_search' })
 export class SavedSearch {
   @PrimaryGeneratedColumn()
@@ -54,7 +98,10 @@ export class SavedSearch {
   @Column({ length: 255, nullable: true })
   name: string;
 
-  @ManyToOne('profile', 'saved_search')
+  @ManyToOne(
+    () => ProfileTable,
+    (profile) => profile.savedSearches,
+  )
   profile: ProfileTable;
 
   @CreateDateColumn()