+++ packages/frontend/tests/stories/languagePicker.spec.ts
@@ -5,6 +5,12 @@ test.describe('Language picker', () => {
   test('Check language picker functionality and query', async ({ page, isMobile }) => {
     await page.goto(defaultAppURL);
 
+    const listItem = page.getByRole('listitem').filter({ hasText: 'Melding om bortkjøring av snø' });
+    const timeElement = listItem.locator('time').filter({
+      hasText: 'Oslo kommune',
+    });
+    await expect(timeElement).toHaveText('Oslo kommune til Test Testesen');
+
     await expect(page.getByTestId('searchbar-input')).toHaveAttribute('placeholder', 'Søk');
     await expect(page.getByRole('button', { name: 'add' })).toHaveText('Legg til filter');
 
@@ -34,6 +40,8 @@ test.describe('Language picker', () => {
 
     await expect(page.getByTestId('searchbar-input')).toHaveAttribute('placeholder', 'Search');
     await expect(page.getByRole('button', { name: 'add' })).toHaveText('Add filter');
+
+    await expect(timeElement).toHaveText('Oslo kommune to Test Testesen');
 
     if (isMobile) {
       await page.getByRole('button', { name: 'Menu' }).click();